Added algorithm and mapbox functionality
The map now shows the vacancy (predicted) for each lot contained in polygons by using red, yellow, and green. MapColorScript when run will update the lot vacancies from the database and print the results into vacancy_data.txt in table format. MapComponent.tsx can read these in when it goes through lotData.json, and using the lot_id when plotting it can match the polygons to the vacancy prediction value and change the color accordingly